+++ jpostman-testng/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.2.3
+
+### Added
+
+* Added automatic TestNG listener registration through Java ServiceLoader.
+* Added `META-INF/services/org.testng.ITestNGListener` to register `JPostmanTestNgAnnotationListener`.
+* Added support for using only `@JPostmanTestNG` on TestNG test classes without also adding `@Listeners(...)`.
+
 ## 1.2.2
 
 ### Added